+++ packages/zenrouter_file_generator/example/lib/routes/(auth)/_layout.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of '_layout.dart';
+
+// **************************************************************************
+// LayoutGenerator
+// **************************************************************************
+
+/// Generated base class for AuthLayout.
+///
+/// URI: /
+/// Path type: stack
+abstract class _$AuthLayout extends AppRoute with RouteLayout<AppRoute> {
+  _$AuthLayout();
+
+  @override
+  NavigationPath<AppRoute> resolvePath(covariant AppCoordinator coordinator) =>
+      coordinator.authPath;
+
+  @override
+  Uri toUri() => Uri.parse('/');
+
+  @override
+  List<Object?> get props => [];
+}