+++ packages/zenrouter_file_generator/example/lib/routes/(auth)/login.g.dart
@@ -1,0 +1,24 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'login.dart';
+
+// **************************************************************************
+// RouteGenerator
+// **************************************************************************
+
+/// Generated base class for LoginRoute.
+///
+/// URI: /login
+/// Layout: AuthLayout
+abstract class _$LoginRoute extends AppRoute {
+  _$LoginRoute();
+
+  @override
+  Type? get layout => AuthLayout;
+
+  @override
+  Uri toUri() => Uri.parse('/login');
+
+  @override
+  List<Object?> get props => [];
+}