+++ packages/zenrouter_file_generator/example/lib/routes/(auth)/register.g.dart
@@ -1,0 +1,24 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'register.dart';
+
+// **************************************************************************
+// RouteGenerator
+// **************************************************************************
+
+/// Generated base class for RegisterRoute.
+///
+/// URI: /register
+/// Layout: AuthLayout
+abstract class _$RegisterRoute extends AppRoute {
+  _$RegisterRoute();
+
+  @override
+  Type? get layout => AuthLayout;
+
+  @override
+  Uri toUri() => Uri.parse('/register');
+
+  @override
+  List<Object?> get props => [];
+}